Separated temporary components/systems header in specific components/systems headers
Linked: #110
